pyproject.toml
[project]
name = "avilla-test"
version = "0.1.0"
description = ""
authors = [
    {name = "Redlnn", email = "w731347477@gmail.com"},
]
dependencies = [
    "aiofile>=3.8.8",
    "aiohttp[speedups]>=3.9.0",
    "aiosqlite>=0.19.0",
    "asyncmy>=0.2.8",
    "avilla-core>=1.0.0a15",
    "avilla-console>=0.2.0",
    "avilla-elizabeth>=1.0.0a15",
    "avilla-qqapi>=1.0.0a19",
    "avilla-qqguild-tencent>=1.0.0a7",
    "avilla-standard-qq>=1.0.0a9",
    "avilla-twilight>=0.2.0",
    "creart>=0.3.0",
    "dnspython>=2.4.2",
    "graia-amnesia>=0.8.2",
    "graia-broadcast>=0.23.3",
    "graia-saya>=0.0.19",
    "graia-scheduler>=0.3.1",
    "graiax-fastapi>=0.4.1",
    "graiax-playwright>=0.3.1",
    "graiax-shortcut>=0.1.0",
    "graiax-text2img-playwright>=0.4.2",
    "jinja2>=3.1.2",
    "kayaku>=0.5.2",
    "launart>=0.8.2",
    "orjson>=3.9.10",
    "pillow>=10.1.0",
    "psutil>=5.9.6",
    "qrcode>=7.4.2",
    "selectolax>=0.3.17",
    "sqlalchemy[asyncio]>=2.0.23",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "AGPL-3.0-only"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm.dev-dependencies]
dev = [
    "black>=23.11.0",
    "isort>=5.12.0",
    "ruff>=0.1.6",
]

[tool.black]
line-length = 120
target-version = ["py310"]
skip-string-normalization = true
safe = true

[tool.isort]
profile = "black"

[tool.ruff]
line-length = 120
target-version = "py311"
select = [
    # https://beta.ruff.rs/docs/rules
    "E",  # pycodestyle
    "F",  # pyflakes
    "UP", # pyupgrade
    "I",  # isort
    # "D",  # pydocstyle
    "ASYNC", # flake8-async
]
exclude = [
    ".git",
    ".venv",
    "__pypackages__",
    "dist",
]
ignore = ["F401"]

[tool.ruff.pydocstyle]
convention = "google"
